Optimizations + some featues
So it turns out all of the png files were at least 1000x1000 pixels, the pawns being 3333x3333 pixels, which is so big that it slowed down the program by a lot.  All of the images are about 128x128 now. Additionally, I added a method to find which square a point is in (line 173 in Form1.cs). One of the more difficult things that I did was add the ability to re-draw a single square, which saves lots of time and doesn't look yucky for lack of a better term.